Relocate RWH Components to be used x-interfaces
Error throw in CI from previous commit as components used in the Support
UI were located in the Candidate UI. In order to maintain consitency I
have extracted these components up a level mirroring the locaion of the
previous WH components.
