+++ app/components/restructured_work_history/gap_component.html.erb
