+++ app/components/restructured_work_history/job_component.html.erb
