+++ app/components/restructured_work_history/work_break_component.html.erb
